Refactor sudoku package into logs and common
The sudoku package is now too large to add any more logic. A new package is going to be introduced, so by this refactor, all packages should be pretty much self contained.

There's now a common package, and also a logs package, which are self-explanatory. Sudoku board is now contained only in the sudoku package.

And finally, a new package will be added called `logic`which will hold all game logic.