Add solver with deterministic algorithm
All logic to solve the board should be contained outside the sudoku package. In case a player wants to game with the board and get hints, on how the complete board looks like.

The solver package will contain all the logic to solve the board, including polling possible values and updating the board.

For now, only add the deterministic algorithm to the solver package. Later on a Backtrack algorithm should be added.