Implement BackTrack Algorithm in solver
Previously backtracking was done in the sudoku file. However, this change moves the logic for backtracking to the solver.

Backtracking is now implemented with availableValues feature. Testing only values that make sense cell by cell, instead of brute forcing with all values. Performance is yet to be tested.

A new package is also introduced as a helper, called version. It holds the required API to keep control of versions of the board, by only keeping the changes in an ordered list.

Finally, logging is changed a bit to denote some logs as debug, and avoid showing them.